atc: structure: turn JobConfig.Plan into function
the goal here is to make it slightly more difficult to misuse the .Plan
field by iterating over it directly instead of recursing through all of
the PlanConfigs contained therein.

the .Plan field has been replaced with .PlanSequence, which on its own
doesn't really help much, but the addition of the .Plan() method which
now returns a PlanConfig (including the job-level hooks) should
hopefully make it a little more obvious that recursion is necessary for
traversing the entire plan config.

note: for the sake of easy skimming, all of the required changes to the
tests have not been included into this commit; they will be committed
separately, and then additional tests will be backfilled in a third
commit.

Signed-off-by: Alex Suraci <asuraci@pivotal.io>