web: behaviour: subscribe to "scrollToId" completion
concourse/concourse#5454

I'd hoped I wouldn't have to do something like this...the issue that
@jomsie discovered has to do with `autoScroll` being `True` when new
events come in, in which case we scroll to the bottom.

One way to alleviate this behaviour without too much effort is to set
`autoScroll` to `False` when we first attempt to `ScrollToID`. This
doesn't work in all cases, though - if we receive the following
messages:

* `EventsReceived` - this is the first time, so we send a `ScrollToID`
message. Suppose the full list of events received in this message is
small enough to fit on screen without scrolling
* `Scrolled` - the port then scrolls to the ID, which triggers this
message. Since the number of log lines received so far is small enough
to fit on the screen, `autoScroll` is set to `True` (since we've
scrolled all the way to the currently-known "bottom" of the page)
* `EventsReceived` - more log lines are received, and since `autoScroll`
was `True`, we `ScrollToBottom`

The only way I could think to get rid of this problem would be to have a
port going from Javascript to Elm. Once we finish the `ScrollToID` from
the JS side, send to the `scrolledToId` port in Elm. During the duration
between starting and finishing the `ScrollToID` call, we'll never
autoscroll (just by keeping track of `isScrollToIdInProgress`).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>